Rename DimensionContentFactory namespace to DimensionContentCollectionFactory
