+++ Content/Application/DimensionContentCollectionFactory/DataMapper/TemplateDataMapper.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ContentBundle\Content\Application\DimensionContentFactory\DataMapper;
+namespace Sulu\Bundle\ContentBundle\Content\Application\DimensionContentCollectionFactory\DataMapper;
 
 use Sulu\Bundle\ContentBundle\Content\Domain\Model\TemplateInterface;
 use Sulu\Component\Content\Metadata\Factory\StructureMetadataFactoryInterface;
