+++ Resources/config/services.xml
@@ -39,7 +39,7 @@
         </service>
 
         <!-- Content Dimension Factory -->
-        <service id="sulu_content.dimension_content_collection_factory" class="Sulu\Bundle\ContentBundle\Content\Application\DimensionContentFactory\DimensionContentCollectionFactory">
+        <service id="sulu_content.dimension_content_collection_factory" class="Sulu\Bundle\ContentBundle\Content\Application\DimensionContentCollectionFactory\DimensionContentCollectionFactory">
             <argument type="service" id="sulu_content.dimension_content_repository"/>
             <argument type="tagged" tag="sulu_content.data_mapper"/>
         </service>
